+++ src/Domain/Entities/Product.java
@@ -5,10 +5,10 @@ import java.util.Scanner;
 public class Product {
     private int id;
     private String name;
-    private int price;
+    private double price;
     public Product(){}
 
-    public Product(int id, String name, int price) {
+    public Product(int id, String name, double price) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -30,11 +30,11 @@ public class Product {
         return name;
     }
 
-    public void setPrice(int _price) {
+    public void setPrice(double _price) {
         this.price = _price;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 }
